convert all routes to require authentication

+++ server/sql/deleteToDo.sql
@@ -1,3 +1,4 @@
 DELETE FROM "ToDo" 
-WHERE "id" = ${toDoId} 
+WHERE "id" = ${toDoId}
+AND creator = ${creator}
 RETURNING *
